Move all code related to the web application to a separate workspace.
