+++ Cargo.toml
@@ -3,4 +3,5 @@
 members = [
     "cli",
     "db",
+    "web",
 ]
