+++ cli/src/main.rs
@@ -2,30 +2,13 @@
 extern crate clap;
 #[macro_use]
 extern crate log;
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate tera;
 
-#[cfg(test)]
-mod firetrack_test;
-#[cfg(test)]
-mod integration_tests;
-
-#[cfg(test)]
-use crate::firetrack_test::*;
-#[cfg(test)]
-use actix_web::test;
-
-mod user;
-
-use actix_files;
-use actix_web::{error, middleware, web, App, Error, HttpResponse, HttpServer};
 use clap::{AppSettings, Arg, SubCommand};
 use db::establish_connection;
 use dotenv;
 use std::env;
 use std::process::exit;
+use web::serve;
 
 static APPLICATION_NAME: &str = "firetrack";
 
@@ -166,90 +149,6 @@ fn main() {
         None => {}
         _ => unreachable!(),
     }
-}
-
-// Starts the web server on the given host address and port.
-fn serve(host: &str, port: &str) {
-    // Configure the application.
-    let app = || {
-        App::new()
-            .wrap(middleware::Logger::default())
-            .configure(app_config)
-    };
-
-    // Start the web server.
-    let addr = format!("{}:{}", host, port);
-    match HttpServer::new(app).bind(addr) {
-        Ok(server) => {
-            server.run().unwrap();
-        }
-        Err(e) => {
-            error!("Failed to start web server on {}:{}", host, port);
-            error!("{}", e.to_string());
-            exit(1);
-        }
-    }
-}
-
-// Controller for the homepage.
-fn index(template: web::Data<tera::Tera>) -> Result<HttpResponse, Error> {
-    let mut context = tera::Context::new();
-    context.insert("title", &"Home");
-    let content = template
-        .render("index.html", &context)
-        .map_err(|_| error::ErrorInternalServerError("Template error"))?;
-    Ok(HttpResponse::Ok().content_type("text/html").body(content))
-}
-
-// Unit tests for the homepage.
-#[test]
-fn test_index() {
-    dotenv::dotenv().ok();
-
-    // Wrap the Tera struct in a HttpRequest and then retrieve it from the request as a Data struct.
-    let tera = compile_templates();
-    let request = test::TestRequest::get().data(tera).to_http_request();
-    let app_data = request.get_app_data().unwrap();
-
-    // Pass the Data struct containing the Tera templates to the index() function. This mimics how
-    // actix-web passes the data to the controller.
-    let controller = index(app_data);
-    let response = test::block_on(controller).unwrap();
-    let body = get_response_body(&response);
-
-    assert_response_ok(&response);
-    assert_header_title(&body, "Home");
-    assert_page_title(&body, "Home");
-    assert_navbar(&body);
-}
-
-// Configure the application.
-fn app_config(config: &mut web::ServiceConfig) {
-    let tera = compile_templates();
-    config.service(
-        web::scope("")
-            .data(tera)
-            .service(actix_files::Files::new("/css", "static/css"))
-            .service(actix_files::Files::new("/images", "static/images"))
-            .service(actix_files::Files::new("/js", "static/js"))
-            .route("/", web::get().to(index))
-            .route("/user/login", web::get().to(user::login_handler))
-            .route("/user/register", web::get().to(user::register_handler))
-            .route("/user/register", web::post().to(user::register_submit)),
-    );
-}
-
-// Compile the Tera templates.
-fn compile_templates() -> tera::Tera {
-    // Determine the path to the templates folder. This depends on whether we are running from the
-    // root of the application (e.g. when launched using `cargo run`) or from the library folder
-    // (e.g. when running tests).
-    let path = if std::env::current_dir().unwrap().ends_with("cli") {
-        "templates/**/*"
-    } else {
-        "cli/templates/**/*"
-    };
-    compile_templates!(path)
 }
 
 // Imports environment variables by reading the .env files.
