+++ web/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cli"
+name = "web"
 version = "0.1.0-dev"
 authors = ["Pieter Frenssen <pieter@frenssen.be>"]
 edition = "2018"
@@ -8,15 +8,11 @@ edition = "2018"
 actix-files = "~0.1"
 actix-http = "~0.2"
 actix-web = "^1.0.8"
-clap = "~2.33"
 db = { path = "../db" }
 dotenv = "^0.15"
-env_logger = "~0.7"
 libxml = "~0.2"
 log = "~0.4"
-regex = "~1.3"
 serde = "~1.0"
 serde_derive = "~1.0"
 tera = "~0.11"
 validator = "~0.10"
-web = { path = "../web" }
