+++ web/templates/base.html
